+++ cmd/validatorinfo.go
@@ -1,4 +1,4 @@
-// Copyright © 2020 Weald Technology Trading
+// Copyright © 2020, 2021 Weald Technology Trading
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -54,7 +54,7 @@ In quiet mode this will return 0 if the validator information can be obtained, o
 		)
 		errCheck(err, "Failed to connect to Ethereum 2 beacon node")
 
-		account, err := validatorInfoAccount()
+		account, err := validatorInfoAccount(ctx, eth2Client)
 		errCheck(err, "Failed to obtain validator account")
 
 		pubKeys := make([]spec.BLSPubKey, 1)
@@ -99,6 +99,12 @@ In quiet mode this will return 0 if the validator information can be obtained, o
 			fmt.Printf("Public key: %#x\n", validator.Validator.PublicKey)
 		}
 		fmt.Printf("Status: %v\n", validator.Status)
+		switch validator.Status {
+		case api.ValidatorStateActiveExiting, api.ValidatorStateActiveSlashed:
+			fmt.Printf("Exit epoch: %d\n", validator.Validator.ExitEpoch)
+		case api.ValidatorStateExitedUnslashed, api.ValidatorStateExitedSlashed:
+			fmt.Printf("Withdrawable epoch: %d\n", validator.Validator.WithdrawableEpoch)
+		}
 		fmt.Printf("Balance: %s\n", string2eth.GWeiToString(uint64(validator.Balance), true))
 		if validator.Status.IsActive() {
 			fmt.Printf("Effective balance: %s\n", string2eth.GWeiToString(uint64(validator.Validator.EffectiveBalance), true))
@@ -107,31 +113,12 @@ In quiet mode this will return 0 if the validator information can be obtained, o
 			fmt.Printf("Withdrawal credentials: %#x\n", validator.Validator.WithdrawalCredentials)
 		}
 
-		//		transition := time.Unix(int64(validatorInfo.TransitionTimestamp), 0)
-		//		transitionPassed := int64(validatorInfo.TransitionTimestamp) <= time.Now().Unix()
-		//		switch validatorInfo.Status {
-		//		case ethpb.ValidatorStatus_DEPOSITED:
-		//			if validatorInfo.TransitionTimestamp != 0 {
-		//				fmt.Printf("Inclusion in chain: %s\n", transition)
-		//			}
-		//		case ethpb.ValidatorStatus_PENDING:
-		//			fmt.Printf("Activation: %s\n", transition)
-		//		case ethpb.ValidatorStatus_EXITING, ethpb.ValidatorStatus_SLASHING:
-		//			fmt.Printf("Attesting finishes: %s\n", transition)
-		//		case ethpb.ValidatorStatus_EXITED:
-		//			if transitionPassed {
-		//				fmt.Printf("Funds withdrawable: Now\n")
-		//			} else {
-		//				fmt.Printf("Funds withdrawable: %s\n", transition)
-		//			}
-		//		}
-
 		os.Exit(_exitSuccess)
 	},
 }
 
 // validatorInfoAccount obtains the account for the validator info command.
-func validatorInfoAccount() (e2wtypes.Account, error) {
+func validatorInfoAccount(ctx context.Context, eth2Client eth2client.Service) (e2wtypes.Account, error) {
 	var account e2wtypes.Account
 	var err error
 	switch {
@@ -147,6 +134,26 @@ func validatorInfoAccount() (e2wtypes.Account, error) {
 		if err != nil {
 			return nil, errors.Wrap(err, fmt.Sprintf("failed to decode public key %s", viper.GetString("pubkey")))
 		}
+		account, err = util.NewScratchAccount(nil, pubKeyBytes)
+		if err != nil {
+			return nil, errors.Wrap(err, fmt.Sprintf("invalid public key %s", viper.GetString("pubkey")))
+		}
+	case viper.GetInt64("index") != -1:
+		validatorsProvider, isValidatorsProvider := eth2Client.(eth2client.ValidatorsProvider)
+		if !isValidatorsProvider {
+			return nil, errors.New("client does not provide validator information")
+		}
+		validators, err := validatorsProvider.Validators(ctx, "head", []spec.ValidatorIndex{
+			spec.ValidatorIndex(viper.GetInt64("index")),
+		})
+		if err != nil {
+			return nil, errors.Wrap(err, "failed to obtain validator information.")
+		}
+		if len(validators) == 0 {
+			return nil, errors.New("unknown validator index")
+		}
+		pubKeyBytes := make([]byte, 48)
+		copy(pubKeyBytes, validators[0].Validator.PublicKey[:])
 		account, err = util.NewScratchAccount(nil, pubKeyBytes)
 		if err != nil {
 			return nil, errors.Wrap(err, fmt.Sprintf("invalid public key %s", viper.GetString("pubkey")))
@@ -212,6 +219,7 @@ func graphData(network string, validatorPubKey []byte) (uint64, spec.Gwei, error
 func init() {
 	validatorCmd.AddCommand(validatorInfoCmd)
 	validatorInfoCmd.Flags().String("pubkey", "", "Public key for which to obtain status")
+	validatorInfoCmd.Flags().Int64("index", -1, "Index for which to obtain status")
 	validatorFlags(validatorInfoCmd)
 }